+++ aletheia/file_types/documents/markdown.py
@@ -9,7 +9,7 @@ class MarkdownFile(PlainTextFile):
     SIGNATURE_WRITE_TEMPLATE = "[//]: # (aletheia:{})"
     SIGNATURE_REGEX = re.compile(
         r'.*\[//\]: # \(aletheia:{'
-        r'"version":(?P<version>' + str(PlainTextFile.SCHEMA_VERSION) + '),'
+        r'"version":(?P<version>\d+),'
         r'"domain":"(?P<domain>[^"]+)",'
         r'"signature":"(?P<signature>[a-f0-9]{2048})"'
         r'}\)$'