+++ tests/file_types/video/test_mp4.py
@@ -71,6 +71,10 @@ class Mp4TestCase(TestCase):
         f = Mp4File(path, cache)
         self.assertRaises(InvalidSignature, f.verify)
 
+    def test_verify_future_version(self):
+        path = self.copy_for_work("future", "mp4")
+        self.assertRaises(UnparseableFileError, Mp4File(path, "").verify)
+
     def test_verify_from_path(self):
 
         path = self.copy_for_work("signed", "mp4")